+++ package/src/components/AudioPlayer/Interface/Controller/Button/index.ts
@@ -1,6 +1,11 @@
 export * from "./PlayBtn";
 export * from "./PrevBtn";
 export * from "./NextBtn";
-export * from "./PlayButton";
+export * from "./TransportControls";
 export * from "./RepeatTypeBtn";
 export * from "./VolumeIcon";
+
+/** @deprecated Use `TransportControls` instead. Will be removed in the next major version. */
+export { TransportControls as PlayButton } from "./TransportControls";
+/** @deprecated Use `TransportControlsProps` instead. Will be removed in the next major version. */
+export type { TransportControlsProps as PlayButtonProps } from "./TransportControls";