+++ package/src/components/AudioPlayer/Interface/Controller/index.tsx
@@ -1,21 +1,23 @@
 import { FC } from "react";
 import "./Controller.css";
 import { useUIContext } from "@/hooks/context/useUIContext";
-import { PlayButton, RepeatTypeBtn } from "./Button";
+import { TransportControls, RepeatTypeBtn } from "./Button";
 import { SortablePlayList } from "./Drawer";
 import { Progress } from "./Input";
 import { Volume } from "./Tooltip";
 
 export const Controller: FC = () => {
   const { activeUI } = useUIContext();
+  const isVisible = (key: keyof typeof activeUI) =>
+    Boolean(activeUI[key] ?? activeUI.all);
 
   return (
     <>
-      <Progress visible={Boolean(activeUI.progress ?? activeUI.all)} />
-      <RepeatTypeBtn visible={Boolean(activeUI.repeatType ?? activeUI.all)} />
-      <PlayButton visible={Boolean(activeUI.playButton ?? activeUI.all)} />
-      <Volume visible={Boolean(activeUI.volume ?? activeUI.all)} />
-      <SortablePlayList visible={Boolean(activeUI.playList ?? activeUI.all)} />
+      <Progress visible={isVisible("progress")} />
+      <RepeatTypeBtn visible={isVisible("repeatType")} />
+      <TransportControls visible={isVisible("playButton")} />
+      <Volume visible={isVisible("volume")} />
+      <SortablePlayList visible={isVisible("playList")} />
     </>
   );
 };